Move KDS to top level
